requirements.txt
fastapi==0.110.0
uvicorn==0.27.1
python-multipart==0.0.12
Pillow==10.2.0
pillow-heif==0.16.0
pydub==0.25.1
pdf2image==1.17.0
yt-dlp>=2025.01.15
aiofiles==23.2.1
requests==2.31.0
